Extract common PM endpoint base
